Crate rosalind_cli

rosalind-cli is the command line interaface for rosalind crate which contains solutions of problems from Rosalind site.

Command structure

rosalind-cli [options]

Options supported by rosalind-cli

General options

-h, --help prints help menu

-t, --task sets task name related to a particular problem

Task specific options

-d, --data sets path to data file

-k offspring amount from each pair (used only in Fibonacci tasks)

-m species lifetime in months (used only in Fibonacci tasks)

-n month amount to calculate population size (used only in Fibonacci tasks)

Problem solutions provided by rosalind-cli

Counting DNA Nucleotides, task name dna

rosalind-cli -t dna -d rosalind_dna.txt

Transcribing DNA into RNA, task name rna

rosalind-cli -t rna -d rosalind_rna.txt

Complementing a Strand of DNA, task name revc

rosalind-cli -t revc -d rosalind_revc.txt

Rabbits and Recurrence Relations, task name fib

rosalind-cli -t fib -n 5 -k 3

Computing GC Content, task name gc

rosalind-cli -t gc -d rosalind_gc.txt

Counting Point Mutations, task name hamm

rosalind-cli -t hamm -d rosalind_hamm.txt

Mendel's First Law, task name iprb

rosalind-cli -t iprb -d rosalind_iprb.txt

Translating RNA into Protein, task name prot

rosalind-cli -t prot -d rosalind_prot.txt

Finding a Motif in DNA, task name subs

rosalind-cli -t subs -d rosalind_subs.txt

Mortal Fibonacci Rabbits, task name fibd

rosalind-cli -t fibd -n 6 -m 3

Inferring mRNA from Protein, task name mrna

rosalind-cli -t mrna -d rosalind_mrna.txt